Add login feature test code
