+++ src/test/java/com/flab/cafeguidebook/service/UserServiceTest.java
@@ -39,4 +39,32 @@ class UserServiceTest {
     );
     assertTrue(isSuccess);
   }
+
+  @Test
+  @DisplayName("이메일, 비밀번호, 이름, 휴대폰번호, 주소, 유저타입이 입력된 경우 회원가입 성공")
+  public void signInTestSuccess(UserDTO user) {
+
+    userService.signIn(user.getEmail(), user.getPassword());
+
+    assertThat(mockHttpSession.getAttribute(SessionKeys.USER_EMAIL)).isNotNull();
+    assertThat(mockHttpSession.getAttribute(SessionKeys.USER_EMAIL)).isEqualTo(user.getEmail());
+
+  }
+
+  @Test
+  @DisplayName("이메일을 잘못 입력시 로그인 실패 및 UserNotFoundException throw")
+  public void signInTestFailWithWrongEmail(UserDTO user) {
+    assertThrows(UserNotFoundException.class, () -> {
+      userService.signIn(user.getEmail() + "Wrong Email", user.getPassword());
+    });
+  }
+
+  @Test
+  @DisplayName("패스워드를 잘못 입력시 로그인 실패 및 UserNotFoundException throw")
+  public void signInTestFailWithWrongPassword(UserDTO user) {
+    assertThrows(UserNotFoundException.class, () -> {
+      userService.signIn(user.getEmail(), user.getPassword() + "Wrong Password");
+    });
+  }
+
 }
